+++ sammoo/__init__.py
@@ -1,4 +1,5 @@
 from .config_selection import ConfigSelection
 from .parmoo_simulation import ParMOOSim
+from .profiles.thermal_load_lpg import ThermalLoadProfileLPG
 
-__all__ = ["ConfigSelection", "ParMOOSim"]
+__all__ = ["ConfigSelection", "ParMOOSim", "ThermalLoadProfileLPG"]